Reorg Kleisli Instances
