+++ kategory/src/main/kotlin/kategory/arrow/Function0.kt
