+++ kategory/src/test/kotlin/kategory/data/KleisliTest.kt
@@ -8,9 +8,9 @@ import org.junit.runner.RunWith
 class KleisliTest : UnitSpec() {
     init {
 
-        val instances = KleisliInstances<Try.F, Int, Throwable>(Try)
+        val me = Kleisli.monadError<Try.F, Int, Throwable>(Try)
 
-        testLaws(MonadErrorLaws.laws(instances, object : Eq<KleisliTKind<Try.F, Int, Int>> {
+        testLaws(MonadErrorLaws.laws(me, object : Eq<KleisliTKind<Try.F, Int, Int>> {
             override fun eqv(a: KleisliTKind<Try.F, Int, Int>, b: KleisliTKind<Try.F, Int, Int>): Boolean =
                 a.ev().run(1) == b.ev().run(1)
